requirements.txt
torch
transformers
datasets
accelerate
peft
bitsandbytes
python-chess
